Instructions
Greetings to everyone who came to look at the Accessus team website developed based on the problem from the Measuring the Value of Earth Observations challenge.
To run the site, you need to clone the repository and have a Node.js to compile the project. 
Enjoy watching! 
If you have any questions, write to the mail albkfil@gmail.com or by phone +77019867333.